Bounty of features and bug fixes  (#94)
* Implement trait constraints in generic code
implement fully qualified method invocations
For this, I did introduce a bit of special syntax. We can remove it later if we are able to make the parser smarter, but it was the shortest path forward for now. If you are referring to a a method via its type, like (Rust) T::clone(), then you must flag that it is a type with a ~. So, in Sway, ~T::clone(). You can see this is used in the implementation of vectors.
This means there are two ways to refer to methods, just like Rust: a.b.foo() and TypeName::foo(a.b), assuming the first argument is self.
Fix some parsing bugs in method invocations
Implement inner struct field reassignments

* remove dbg print

* remove infinite recursive dependency

* fix up sway-server for new reassignment type

* fix warnings

* cargo fmt

* sketch out asm for struct subfield reassignment

* middle of airport progress...

* add struct memory descriptor abstraction for use in code generation

* clean up todos; finish struct inner field reassignment assembly

* add struct inner field reassignment to test suite

* add test for struct memory layout

* fix warnings

* fix tests

* remove dead comment and fix formatting

Co-authored-by: Alexander Hansen <alexanderhansen@Alexanders-MBP.localdomain>
Co-authored-by: Alexander Hansen <alexanderhansen@Alexanders-MacBook-Pro.local>