Replace the CompileResult enum with a struct.
The CompileResult was a bit redundant in that both the arms for Ok and
Err had `warnings` and `errors` vectors.  Frequently when matching a
CompileResult the `warnings` and `errors` would be copied redundantly,
regardless of the actual result value.

By replacing the enum with a struct which has one copy of warnings and
errors and an Option<T> for the value instead we can reduce boilerplate
considerably.  This commit has a net change of -400 lines of code from
the Sway codebase.