Detect and report errors for recursive function calls. (#221)
The dependency analysis now uses a HashMap of HashSets instead of a 2D
matrix of &str.  Each dependency is to and from a DependentSymbol
wrapper which now special cases function declarations.

With the function declarations and their dependencies gathered it's
relatively easy to check whether any functions (eventually) depend on
themselves.  New errors are now reported when recursion is detected.