Fix bug in asm generation related to LWDataId virtual op.
Previously when replacing labels with offsets it assumed that all
virtual ops are 1 word in size after conversion to VM ops.

This isn't necessarily true for LWDataId which can be realised as an LW
followed by an ADD when loading a data section offset, for datatypes
larger than a single word.