Change the way trait implementations are handled by the namespace.
We now insert trait implementations into the namespace where they're
declared, rather than where the trait is declared.  This involves using
a full callpath to the trait as the key for the trait lookup hashmap.

We therefore also lookup trait implementations in the local namespace to
resolve them.

Unfortunately this breaks trait implementations for standard types, like
those found in `std::ops` for `std::ops::Ord` because without a fully
qualified path to the types we will look for their trait implementations
locally and not find them.

A hack has been added to the lookup function which will first try the
local namespace and then if unsuccessful try the module where the trait
is declared.  This works for all our current tests, but is fragile and
it's bad to assume that we'll always find implementations for standard
types where the trait is declared.