Add a constant combining pass.
This first version focuses on `insert_value` when both the struct and
the value inserted are constant.  The code in Sway (not yet merged as of
writing) creates a lot of these as it builds constant structs up from
scratch, one field at a time.   This pass streamlines that process of
creating const structs.

See in particular the new `constants_insert_value.*` tests for an
example.