Sway IR crate. (#596)
* Introduce the source for the Sway IR crate.

* Add sway-ir to Cargo and give it a README.

* Format the README a bit.

Co-authored-by: John Adler <adlerjohn@users.noreply.github.com>

* Bump version to 0.2.1.

Co-authored-by: John Adler <adlerjohn@users.noreply.github.com>

* Move the inlining code into its own module.

* Add a constant combining pass.

This first version focuses on `insert_value` when both the struct and
the value inserted are constant.  The code in Sway (not yet merged as of
writing) creates a lot of these as it builds constant structs up from
scratch, one field at a time.   This pass streamlines that process of
creating const structs.

See in particular the new `constants_insert_value.*` tests for an
example.

* Add rustdocs to pretty much everything.

Basic descriptions at this stage, no example code nor doc tests yet.

Co-authored-by: John Adler <adlerjohn@users.noreply.github.com>