Use `Ident` for the strings in `AsmBlock`s.
This change means the IR asm block closely resembles the Sway asm block,
and so the code generation tools which already exist in `sway-core` can
be easily re-used.

`Ident` also contains a `Span` which is another step in the right
direction, as we need to add span info to all IR instructions soon.

So eventually all identifiers in the IR should switch from `String` to
`Ident`; this is the first step.