Optional IR integration into `sway-core`. (#640)
* Add IR test files.

* Add command line options supporting IR to `forc`.

- `forc build --use-ir` to choose to use the IR pipeline.
- `forc build --print-ir` to print the IR instead, much like the other `--print-*` options.

* Some small updates to the E2E tests.

In particular the dynamic OOB test is disabled as we won't support
runtime bounds checks any longer.  BUT!  This test could be found at
compile time with just the right IR analysis, so it's worth keeping
until we can re-enable it.

* Use `Ident` for the strings in `AsmBlock`s.

This change means the IR asm block closely resembles the Sway asm block,
and so the code generation tools which already exist in `sway-core` can
be easily re-used.

`Ident` also contains a `Span` which is another step in the right
direction, as we need to add span info to all IR instructions soon.

So eventually all identifiers in the IR should switch from `String` to
`Ident`; this is the first step.

* Add support for compiling via IR.