sway-core: Expose inter-compilation steps for use in forc
This refactors the bodies of the `compile_to_asm` and
`compile_to_bytecode` functions into `ast_to_asm` and `asm_to_bytecode`
respectively. This allows `forc` to compile the AST (for producing the
ABI JSON) and then re-use the compiled AST to complete the rest of
compilation to bytecode in the same pass.