Construct `pkg_graph` at beginning of `forc build`
This is a start at separating the construction of the dependency graph
and fetching of dependency source from the compilation process. This is
done under a new `fetch_pkgs` function where, given the manifest of the
package we're compiling, it produces a graph where each node represents
a package at a known, pinned version/commit along with a `Path` to a
local copy of their source (under `.forc/git/checkouts` for git deps).

The `pkg_graph` also contains our root-level package, allowing us to
treat all packages in the compilation process as equal, avoiding special
case code for the top-level program.

The compilation order of packages is determined by performing a toposort
of the `pkg_graph`. By compiling packages in this order, each node is
always guaranteed to have access to the compiled artifacts of each of
its dependencies.

This PR also introduces the `git2` crate for fetching and working with
git dependencies. This should allow us to remove the current
github-specific dependency handling in favour of more general git
support.

---

I think the next step in this PR will be replacing the existing
`dependency_graph` mutable hash map reference with an immutable
reference to the newly constructed `pkg_graph`. It seems this may also
involve populating the `Namespace` prior to the rest of compilation too?
I'm still familiarising myself with `sway-core`, so will need to dive a
bit deeper to be sure.