Remove the unused `&mut dependency_graph` throughout `sway-core`
While working on #825 I noticed that that the `dependency_graph` is
unused throughout `sway-core`. I think the reason why cargo doesn't emit
any warnings about this is that at one point, it's stored in the public
`TypeCheckArgument` struct, so cargo believes that it is intentionally
being re-exported, however in reality we don't use it anywhere else in
the fuel/sway/forc ecosystem.