Smaller IR changes. (#828)
* Introduce IrError and use it for most Results in the library.

* Handle IrErrors naively by converting to String.

* Remove the redundant Module name.

I put it there initially without thinking, and then just used the name
'script' for scripts, etc.  This just removes it altogether; touches a
lot of the test files.

* Move the aggregate naming management from IR back into core.

The IR doesn't care about struct field names nor enum variant names.
They need to be tracked by IRgen though, so instead of managing it in
the IR Context it can be all done in IRgen core.