Improve support for determinism and cachability with `Forc.lock`
This adds a `Forc.lock` file that for the most part mirrors the
`Cargo.lock` file.

If a `Forc.lock` doesn't exist, a `Lock` structure is created from the
package graph and serialised to TOML before being written to the
`Forc.lock` file path.

If a `Forc.lock` *does* exist, we construct the build graph from the
`Lock` structure, re-using the previously fetched source for the pinned
dependencies.

TODO:

- Make `forc update` update the lock file.
- Add support for updating individual packages.