Switch to _only_ calling inlined generated functions for IR.
Before now when compiling a function call to IR, if the named callee
could be found in the already compiled function list then we'd call
that.  Otherwise we'd recreate the callee, name it `anon_X` and call
that.

Unfortunately our name resolution isn't rich enough and it was too easy
to get name clashes and/or ambiguity, and it was possible for the wrong
function to be called (especially trait methods, which have a consistent
interface).

So, for now, until we can use fully qualified absolute paths for
function names, all calls are to recreated functions.  We only compile
`main()` and the functions inlined to it.