Big old bag of IR fixes. (#852)
* Be sure to resolve the Self type for ImplSelf decls during semantic analysis.

* Add a test for IR inlining not working properly.

The inlined function must have an `if` expression and be called (i.e.
inlined) twice, exercising a different path through the `if` arms.

* Stop caching const values in IR -> ASMgen.

I put a big comment in there explaining why.

* Remove a wayward dbg!() call.

* Add a test and fix for an IR bug WRT methods on empty structs.

* Fix for suggestions from the new Clippy.

* Switch to _only_ calling inlined generated functions for IR.

Before now when compiling a function call to IR, if the named callee
could be found in the already compiled function list then we'd call
that.  Otherwise we'd recreate the callee, name it `anon_X` and call
that.

Unfortunately our name resolution isn't rich enough and it was too easy
to get name clashes and/or ambiguity, and it was possible for the wrong
function to be called (especially trait methods, which have a consistent
interface).

So, for now, until we can use fully qualified absolute paths for
function names, all calls are to recreated functions.  We only compile
`main()` and the functions inlined to it.

* Fix the IR parser to allow out of order function declarations.

Surprisingly, until now all the tests have passed while the parser
required function declarations to be in dependency order.  Now, while
we're generating `anon_X()` functions for every call in `main()` they
are no longer in order.

For the parser to recreate the IR context from source it now does two
passes; the first creates all functions while leaving placeholders for
CALLs and the second will resolve those placeholders once all functions
have been created.

This means we now have a NOP instruction for the placeholder, and a new
Block method for replacing specific individual instructions.

* More Clippy fixes.