Move forc's util and pkg related code into `forc-util`, `forc-pkg` (#901)
* Move forc's util and pkg related code into `forc-util`, `forc-pkg`

This makes a start on #863 by refactoring the "package" handling code
into a `forc-pkg` crate. This new crate is solely responsible for
building, fetching, locking and updating of forc packages.

`forc-pkg` pulls in around 170 dependencies altogether, as opposed to
`forc`'s current ~460. This should be useful for downstream crates that
may want to depend on `forc` without all the bloat of the CLI tooling,
explorer, etc. The `sway-lsp` crate comes to mind.

To achieve this, some of forc's utility functionality has been moved
into `forc-util` so that it may be shared between `forc-pkg` and `forc`
itself.

* Address some clippy nits

* Update tests for parsing the default manifest and test manifest strings

* Fix dependency order in forc-pkg Cargo.toml

* Improve `forc-pkg`, `forc-util` manifest descriptions

Also tweaks the version declarations for forc-pkg dependencies.

* Specify full version for sway deps to be publish-friendly

* Specify full versions for `forc-*` deps ready for publishing