More IR updates. (#1034)
* Fix IR for E2E tests... again.

- Fix for new RETD behaviour.
- Remove the 'Register' storage mode in ASMgen, as it was buggy and
  performing an optimisation which should be done back in IR passes.

* Change the 'data section start' register from 46 to 63.

The VM has 64 registers.  The first 16 are reserved, the following 48
are general purpose.  The compiler reserves one of the general purpose
registers to always contain the address of the data section.

Until now it was chosen as 'the number of general purpose registers
minus 2' which is 48 - 2 = 46.  This is in the middle of those used by
the compiler and prone to clobbering.  Now it is 'the total number of
registers minus 1' AKA 'the last general purpose register', i.e., 63.

Also, the `fix_opcode_bug` test uncovered this bug, and happened to be
misnamed in its `Forc.toml` file.

* Introduce and use `sway_ir::Type::is_copy_type()`.