Improve support for git reference kinds (`branch`, `tag`, `rev`), improve fetching behaviour (#1045)
* Add proper support for git reference kinds (`branch`, `tag`, `rev`)

Works toward resolving #1019, however tests and examples still need to
be updated.

Previously, we initialised the temporary git repository with a `clone`
operation. By default, this only fetches a subset of git references,
meaning that using references other than those provided by default with
`clone` failed to resolve.

Inspired by cargo, this commit updates the behaviour to instead first
initialise the repository, then attempt to fetch only those references
that are relevant to the user specified git reference.

In order to construct an accurate set of `refspecs` and fetch only the
relevant commits, its useful to retain knowledge about the *kind* of git
reference. This introduces a new `GitReference` type and updates the git
source lock file serialization to account for this.

* Update example `Forc.lock` files for git reference changes

* Update E2E test `Forc.lock` files for git reference changes

* Validate the commit hash while parsing `SourceGitPinned` from str