Improve strings for IR. (#1113)
* Improve IR printer.

Print any const args to `extract_value` and only print const values once
ever.

* Change the IR syntax for unions.

It used to be `{ ty0 | ty1 | ... }` but a single variant union would
just be `{ ty }`.  This is indistinguishable from a single field struct
or tuple so now it's `( ty0 | ty1 | ... )`.

* Improve string support in the IR.

Assume that strings are always used by reference and so their stack size
will always be that of one word.

IR treats strings as a vector of bytes to avoid UTF-8 trickiness.  They
may still be UTF-8 strings of course, but the IR doesn't care either
way.