Allow forwarding options and arguments from `forc test` to `cargo test`
This allows for forwarding both custom options and custom arguments
through to `forc test`'s underlying `cargo test` invocation.

See the documentation for the new `forc_test::Command` fields for
information on how to forward options and arguments.

This also removes the permanent "--nocapture" arg that we previously
passed through to `cargo test` in favour of allowing the user to
optionally specify it if they wish.

To achieve the original behaviour, you can invoke `forc test` with the
following:

```
forc test -- --nocapture
```

Note that this PR *does* keep the original behaviour of permanently
setting `--color always` in order to provide nice colorful output by
default. This is useful as cargo appears to disable colored output by
default when run as a child process.

In the case that the user wishes to disable colored output, they can
still do so by specifying the necessary `cargo test` option, e.g.

```
forc test --cargo-test-opts="--color never"
```

Closes #1004