WIP: Remove global namespace arena in favour of structural sharing
Closes #682

This removes the global arena used for managing namespaces during type
checking in favour of reverting to the original approach of representing
new scopes by cloning the namespace.

Prior to the introduction of the global arena, cloning the namespace to
represent new scopes caused exponential memory growth as the entirety of
each of the inner collections was cloned along with the namespace
itself.

In this new approach, we aim to avoid this by using clone-friendly,
drop-in-replacement collections from the `im` crate that enable
"structural sharing" of the `Namespace` fields. *See TODO below for
whether this is done yet or not.*

You can read more about structural sharing [here][1]. The gist is that
rather than deep cloning the entire structure, we clone a reference to
the original state, and new allocations only occur upon *modifying* the
data structure with only the memory needed to record the difference.

The result is that we can avoid the side-effecting nature of maintaining
global state, making it easier to reason about how `Namespace`s are
accessed and updated within functions simply by reading their
signatures.

It should also allow us to lean on RAII for cleaning up the namespace
contents, solving the bug where we are currently unable to compile the
same project more than once within the same process as mentioned
[here][2]. This is particularly relevant to `sway-lsp` where we will
likely want to lean on `sway-core` to type check the same code many
times within one process.

This also gets us a step toward achieving referential transparency for a
lot of our type checking functions which may enable easier caching and
parallelism in the future, though this would still currently be blocked
by the type engine's global state.

Many of the newer methods previously provided by `NamespaceWrapper` and
`NamespaceRef` have been moved to `Namespace` itself.

[1]: https://docs.rs/im/latest/im/#what-are-immutable-data-structures
[2]: https://github.com/FuelLabs/sway/issues/682#issuecomment-1039907300

TODO
----

- [ ] Solve E2E errors. Since the arena was introduced, a lot of new
  global mutation has also been introduced disguised by the `RwLock`.
  E.g. the `monomorphize` functions now mutate the namespace at the
  given reference, meaning that `ns.resolve_self_*` is also
  self-mutating. Need to review how these methods are used to ensure the
  new behaviour achieves the same intended results.
- [ ] Replace collections in `Namespace` fields with clone-friendly
  alternatives provided by `im`.