Introduce a path-aware `ManifestFile` type (#1249)
* Introduce a path-aware `ManifestFile` type

After loading a `Manifest`, most useful operations involving it require
knowledge of the directory path from which it was loaded from (i.e.
`entry_string`, `program_type`).

This introduces a `ManifestFile` type that stores the path to the
`Forc.toml` from which its inner `Manifest` was read. This allows to
avoid a lot of unnecessary code duplication related to traversing the
parents to find the root manifest directory. Now, we can just load the
`ManifestFile` and call `manifest.path()` or `manifest.dir()` to get the
path to the `Forc.toml` or its enclosing directory respectively.

* Use correct manifest directory path for `fetch_deps`

* Replace comment previously lost in rebase conflict resolution

* Use existing error constructor provided by pkg module