Handle removed dep in manifest during path reconstruction from graph
This fixes a bug where removing a path dependency (e.g. core) from the
manifest where that dependency is shared by another dependency (e.g.
std) could cause path map reconstruction to fail.

Previously, we assumed that the dependency would be present within the
manifest in order to retrieve the source path. This PR removes this
assumption and will instead reconstruct the dependency graph (and in
turn the lock file) when a dependency is discovered to be missing during
path reconstruction.

First spotted by @nfurfaro [here][1].

[1]: https://github.com/FuelLabs/sway/pull/1276/files#r853131667