Refactor namespaces items into ergonomic `Namespace` session type
This refactors the `init`, `root` and `mod_path` namespace items into a
single `Namespace` context type. This simplifies passing these fields
through `TypeCheckArguments` and the rest of type-checking and abstracts
away some of the complexity involved in interactions between each of
these items.

For the most part, the new `Namespace` methods are just short-hand for
using `root` and `mod_path` together, however there are a couple of
significant additions:

- `Namespace::enter_submodule` returns a session type with a namespace
  associated with the submodule being type-checked. Upon being dropped,
  the `SubmoduleNamespace` returns control to the current module and
  resets the associated `mod_path`.
- `Namespace::enter_scope` simply clones each of the fields, but better
  clarifies intentions in code.