Re-add `compile_inner_dependency` to root
Originally I inlined this into the `import_new_file` as it required some
refactoring following changes to the `Namespace`, and was the only
use site anyway. Thinking on this more, it makes sense to abstract this
step to the root as currently, it involves both parsing and
type-checking, not just type-checking.