Implement `Display` for `TypeId`s;  Add type parameters to the definition of custom types (#1534)
Include type_parameters in the definition of a struct or enum in the type engine. This allows for handling of generic types where the type is not necessarily used in the fields or variants.
Change TypeId from a type alias to a newtype. This provides greater type safety but also allowed me to implement Display for it. Now when you print a TypeId, instead of the numeric value in the type engine, you'll see the actual type it represents.
Do a bit of refactoring/code cleanup in places.