Ensure lock is applied when `BuildPlan` validation would require changes (#2090)
Ensure lock is applied if BuildPlan validation requires changes

While reviewing #2085, I noticed that some recent additions to
`BuildPlan` validation could result in some changes to the build plan
(and in turn, rewriting the lock file) even in the case that `--locked`
was specified.

This moves the `--locked` check to the moment before writing the new
lock file. This ensures all potential changes that would be required of
the `BuildPlan` are caught. It also allows us to provide a "Cause" for
*why* the lock file would need updating when `--locked` is passed to
prevent it.

Closes #2084
Closes #2085