`forc-pkg` - Don't add transitive deps (besides `core`) to namespace
Currently `forc-pkg` adds all transitive dependencies to a package's
initial namespace prior to its compilation. This had the benefit of
implicitly including `core`, but with the downside of implicitly
including every other transitive dependency package, even if not a
direct depednency (not what we want).

This PR changes the behaviour to only include direct dependencies and
`core` within a package's initial namespace.

Addresses 1, 2 of #2125.