Remove duplicate CI checks that already have dedicated jobs
I noticed a bunch of our CI jobs were were duplicating existing checks.
In particular, there was a lot of copy-paste steps of installing forc
and the forc-fmt plugin, even when unused in the following checks.

This doesn't improve the real bottleneck (our stdlib test job), but it's
a start.