Check for calling of associated functions as method calls.
This adds a new error and checking for calls of associated functions
using method call syntax.

This means for code like the following:

```
struct Bar {}

impl Bar {
  fn associated() {}
}

fn main() -> u64 {
  let bar = Bar {};
  bar.associated();
  0
}
```

we now emit an error:

```
error
  --> main.sw:11:3
   |
 9 |
10 |   let bar = Bar {};
11 |   bar.associated();
|   ^^^^^^^^^^^^^^^^ Cannot call associated function "associated" as
method call. Use associated function syntax instead.
12 |   0
13 | }
   |
```

This brings us closer to Rust behaviour, which also checks for this and
only allows functions taking a self to be called with method call
syntax.

It adds a few more tests for this and also updates some standard library
trait methods to take self so they work with the new check.