Minimise I/O by replacing BuildProfile's PathMap with a ManifestMap
This commit reduces the number of times that we need to read each
dependency's manifest from file in order to reduce the amount of I/O
performed. Each dependency's manifest should now only be loaded once
during either `graph_to_manifest_map` or `fetch_graph`.

This required implementing `Clone` for `ManifestFile` to allow for
constructing the build plan in a manner that does not require consuming
the manifest file. `ManifestFile::clone` should only ever be called for
the root node.

As a side-effect, the `build` and `check` commands no longer require
loading each package's manifest as they are already provided by the
build plan. As a result both commands no longer require the sway git
tag to be passed through.